Update mqtt_server.c
Socket connections seem to be sticky and peer abortion is not detected by the server.

Ran into problems with following configuration: #1. ESP32 with server, publisher and subscriber. #2. ESP32 with publisher and subscriber. #3 Generic MQTT client on android device subscribing and publishing. 

Number of allowed sockets is limited which creates problem in #1. Socket connections do not close on sudden abortion of #2 (reset/out of range). When #2 tries to connect again it obtains a new connection without previous connection being deleted. Result: socket handler gets saturated and returns error. Socket number is limited to 16 and I find no way around that.

Proposed fix: revival of clientid registration code (muted) in mqtt_server.c. On new connection: check if there is a working connection with the same clientid and force that one to close  if so.

This fix has been tested and working.